Dynamic title tags + php rewrite of index page
Every page now has the possibility for a dynamic title. The header file
checks whether variable $title is set in the file it has been called by
(e.g. nieuws.php). If so, it sets the title to '$title | Homemade Water'
(where $title is replaced by its content).

The php part where the agenda and news items are called on the homepage,
is rewritten as well. This is mainly done because of indentation
purposes. In a similar way this is done for most of the other pages as
well. (note: the tab option in dreamweaver gives a rather nasty result
on the returned html page to the server).