Added a gastenboek (guest book) in its early stage
A gastenboek is now present on the site as well (development)
-There is another table on the website in which all the posts are saved.
- The reading of which posts are done and putting those on the webpage
works in BETA (it takes all posts for now, restriction on that is
needed, and then a previous & next button would be nice)
- The form on top, through which one can upload its post should still be
fixed (css as well as the proccessing of the input. Probably it's best
to put the whole gastenboek in a new db, because of restrictions and
such.

+ late upload of photoalbum javascript changes (whoops)